fix(test): enhance test isolation and remove deprecated files
- Revised TEST_ISOLATION_ANALYSIS.md to reflect the successful resolution of all test isolation issues, achieving a 96.3% test success rate.
- Improved TESTING_GUIDE.md with updated statistics and documentation on the elimination of test isolation problems.
- Deleted outdated collision test files and refactored the UI Manager and collision utility files to streamline the codebase and enhance modularity.

This commit is part of the ongoing effort to improve test reliability and maintainability across the project.
